refactor(api): move reset scorecard route to evluation scorecard context
