+++ api/src/evaluation/application/scorecards/index.js
@@ -1,3 +1,7 @@
+import Joi from 'joi';
+
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { scorecardController } from './scorecard-controller.js';
 
 const register = async function (server) {
@@ -26,6 +30,32 @@ const register = async function (server) {
             '- L’id demandé doit correspondre à celui de l’utilisateur authentifié',
         ],
         tags: ['api'],
+      },
+    },
+    {
+      method: 'POST',
+      path: '/api/users/{userId}/competences/{competenceId}/reset',
+      config: {
+        pre: [
+          {
+            method: securityPreHandlers.checkRequestedUserIsAuthenticatedUser,
+            assign: 'requestedUserIsAuthenticatedUser',
+          },
+        ],
+        validate: {
+          params: Joi.object({
+            userId: identifiersType.userId,
+            competenceId: identifiersType.competenceId,
+          }),
+        },
+        handler: scorecardController.resetScorecard,
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+            "- Cette route réinitialise le niveau d'un utilisateur donné (**userId**) pour une compétence donnée (**competenceId**)",
+          '- Cette route retourne les nouvelles informations de niveau de la compétence',
+          '- L’id demandé doit correspondre à celui de l’utilisateur authentifié',
+        ],
+        tags: ['api', 'user', 'scorecard'],
       },
     },
   ]);
